requirements.txt
pygame-ce==2.5.6
pygame_gui==0.6.14